backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pdfplumber==0.11.4
langchain>=1.0.0
langchain-google-genai>=4.0.0
langchain-core>=1.0.0
python-dotenv==1.0.1
pydantic==2.9.2
